fix(server): include workflow-runtime work in operator-snapshot
Merge aged nonterminal and failed workflow instances into
retry.stalled_tasks and recent_failures so runtime-only submissions are
visible beside TaskStore rows, without double-counting shared handles.
